+++ emp-tool/emp-tool/utils/hash.h
@@ -65,7 +65,7 @@ class Hash { public:
 	}
 
 	static block KDF(Point &in, uint64_t id = 1) {
-		size_t len = in.size();
+		uint64_t len = in.size();
 		in.group->resize_scratch(len+8);
 		unsigned char * tmp = in.group->scratch;
 		in.to_bin(tmp, len);
